# exchange-gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Exchange API
  version: v1.0.0
  x-contract-id: exchange-gateway
  description: |
    Cryptocurrency Exchange Platform API

    ## Authentication

    Private endpoints require HMAC-SHA256 signature authentication:

    **Required Headers:**
    - `X-API-KEY`: Your API Key
    - `X-API-TIMESTAMP`: Unix timestamp in milliseconds
    - `X-API-NONCE`: Random string (UUID recommended)
    - `X-API-SIGNATURE`: HMAC-SHA256 signature

    **Signature Calculation:**
    ```
    canonicalQuery = sorted query string (exclude `signature` if present)
    canonical = timestamp + "\n" + nonce + "\n" + METHOD + "\n" + path + "\n" + canonicalQuery
    signature = HMAC-SHA256(apiSecret, canonical)
    ```
    Note: request body is not included in the signature.

    **Permissions:**
    - `1` = Read (query orders, balances)
    - `2` = Trade (place/cancel orders)
    - `4` = Withdraw (request withdrawals)

    ## Rate Limits

    - Public endpoints: 1200 requests per minute per IP
    - Private endpoints: 600 requests per minute per API key
    - Orders: 10 requests per second per API key

    When rate limited, the API returns HTTP `429` with `RATE_LIMIT_EXCEEDED`.
    Response headers include:
    - `X-RateLimit-Limit`
    - `X-RateLimit-Remaining`
    - `X-RateLimit-Reset` (Unix timestamp in seconds)

    ## Error Codes

    Common error codes returned in `ErrorResponse.code`:
    - `INVALID_SIGNATURE`: Signature mismatch
...

    ## WebSocket

...

...

...

  contact:
...
servers:
...
paths:
...
  /v1/order:
    post:
      tags: [Trading]
      summary: Place Order
      description: Create a new order (requires Trade permission)
      operationId: createOrder
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              symbol: BTCUSDT
              side: BUY
              type: LIMIT
              timeInForce: GTC
              price: 4500000000000
              quantity: 100000000
              clientOrderId: "my-order-001"
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              example:
                orderId: 1735012345678001
                clientOrderId: my-order-001
                symbol: BTCUSDT
                side: BUY
                type: LIMIT
                timeInForce: GTC
                price: "45000.00000000"
                origQty: "1.00000000"
                executedQty: "0.00000000"
                status: NEW
                createdAt: 1703232000000
                updatedAt: 1703232000000
        '400':
          description: Invalid order parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INVALID_ORDER
                message: Price is required for LIMIT orders
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INVALID_SIGNATURE
                message: Signature mismatch
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: RATE_LIMIT_EXCEEDED
                message: Too many requests

    get:
      tags: [Trading]
      summary: Query Order
      description: Get order details by orderId
      operationId: getOrder
      security:
        - ApiKeyAuth: []
      parameters:
        - name: orderId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: Order ID
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: ORDER_NOT_FOUND
                message: Order does not exist

    delete:
      tags: [Trading]
      summary: Cancel Order
      description: Cancel an active order
      operationId: cancelOrder
      security:
        - ApiKeyAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: orderId
          in: query
          schema:
            type: integer
            format: int64
          description: Order ID (either orderId or clientOrderId required)
        - name: clientOrderId
          in: query
          schema:
            type: string
          description: Client order ID
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: ORDER_NOT_CANCELABLE
                message: Order already filled
  /v1/openOrders:
    get:
      tags: [Trading]
      summary: Current Open Orders
      description: Get all open orders for a symbol or all symbols
      operationId: getOpenOrders
      security:
        - ApiKeyAuth: []
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
          description: Filter by symbol (optional)
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        '200':
          description: List of open orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /v1/allOrders:
    get:
      tags: [Trading]
      summary: All Orders
      description: Get all orders (active, canceled, filled)
      operationId: getAllOrders
      security:
        - ApiKeyAuth: []
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: integer
            format: int64
          description: Start time (Unix ms)
        - name: endTime
          in: query
          schema:
            type: integer
            format: int64
          description: End time (Unix ms)
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /v1/myTrades:
    get:
      tags: [Trading]
      summary: Account Trade List
      description: Get trades for a specific account and symbol
      operationId: getMyTrades
      security:
        - ApiKeyAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: startTime
          in: query
          schema:
            type: integer
            format: int64
        - name: endTime
          in: query
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
      responses:
        '200':
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountTrade'

  # ==================== Account Endpoints ====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: |
        HMAC-SHA256 authentication. Required headers:
        - X-API-KEY: Your API key
        - X-API-TIMESTAMP: Unix timestamp (ms)
        - X-API-NONCE: Random UUID
        - X-API-SIGNATURE: HMAC-SHA256 signature
    BearerAuth:
      type: http
      scheme: bearer
      description: "Bearer token from login (format: v1.<payload>.<signature>)"
  schemas:
...
    CreateOrderRequest:
      type: object
      required: [symbol, side, type, quantity]
      properties:
        symbol:
          type: string
          example: BTCUSDT
        side:
          type: string
          enum: [BUY, SELL]
        type:
          type: string
          enum: [LIMIT, MARKET]
        timeInForce:
          type: string
          enum: [GTC, IOC, FOK]
          default: GTC
          description: Good-Til-Canceled, Immediate-Or-Cancel, Fill-Or-Kill
        price:
          type: integer
          format: int64
          description: Price in smallest unit (required for LIMIT orders)
          example: 4500000000000
        quantity:
          type: integer
          format: int64
          description: Quantity in smallest unit
          example: 100000000
        quoteOrderQty:
          type: integer
          format: int64
          description: Quote quantity for MARKET orders
        clientOrderId:
          type: string
          maxLength: 36
          description: Client-defined order ID
          example: "my-order-001"
...
    Order:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          example: 1735012345678001
        clientOrderId:
          type: string
        symbol:
          type: string
          example: BTCUSDT
        side:
          type: string
          enum: [BUY, SELL]
        type:
          type: string
          enum: [LIMIT, MARKET]
        timeInForce:
          type: string
          enum: [GTC, IOC, FOK]
        price:
          type: string
          example: "45000.00000000"
        origQty:
          type: string
          example: "1.00000000"
        executedQty:
          type: string
          example: "0.50000000"
        status:
          type: string
          enum: [NEW, PARTIALLY_FILLED, FILLED, CANCELED, REJECTED]
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
    AccountTrade:
      type: object
      properties:
        tradeId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        symbol:
          type: string
        price:
          type: string
        qty:
          type: string
        commission:
          type: string
        commissionAsset:
          type: string
        time:
          type: integer
          format: int64
        isBuyer:
          type: boolean
        isMaker:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: INVALID_SYMBOL
        message:
          type: string
          example: "Symbol XXXUSDT is not valid"